refactor: News 정보 수집 Event listener 패키지로 변경
