+++ src/main/java/com/hanium/mom4u/domain/news/service/NewsScheduler.java
@@ -1,5 +1,6 @@
 package com.hanium.mom4u.domain.news.service;
 
+import com.hanium.mom4u.domain.news.listener.S3JsonImportEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
