+++ src/main/java/com/hanium/mom4u/global/crawling/controller/CrawlTestController.java
@@ -1,6 +1,6 @@
 package com.hanium.mom4u.global.crawling.controller;
 
-import com.hanium.mom4u.domain.news.service.S3JsonImportEvent;
+import com.hanium.mom4u.domain.news.listener.S3JsonImportEvent;
 import com.hanium.mom4u.external.s3.FileStorageService;
 import com.hanium.mom4u.global.crawling.service.CrawlingTriggerService;
 import lombok.RequiredArgsConstructor;
